renamed stream.rs to collection.rs
